chore/Corrected previous refactor, I failed to save the previous changes
