+++ src/components/banner-section/banner-gradient/BannerGradient.tsx
