+++ src/components/banner-section/banner-image-container/BannerImageContainer.tsx
