+++ src/stories/components/banner-section/BannerGradient.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { BannerGradient } from '@/components/banner/banner-gradient/BannerGradient';
+import { BannerGradient } from '@/components/banner-section/banner-gradient/BannerGradient';
 
 const meta: Meta<typeof BannerGradient> = {
   component: BannerGradient,
-  title: 'Components/Banner/BannerGradient',
+  title: 'Components/BannerSection/BannerGradient',
   argTypes: {},
 };
