+++ src/stories/components/banner-section/BannerImageContainer.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { BannerImageContainer } from '@/components/banner/banner-image-container/BannerImageContainer';
+import { BannerImageContainer } from '@/components/banner-section/banner-image-container/BannerImageContainer';
 
 const meta: Meta<typeof BannerImageContainer> = {
   component: BannerImageContainer,
-  title: 'Components/Banner/BannerImageContainer',
+  title: 'Components/BannerSection/BannerImageContainer',
 };
 
 export default meta;
